// src/app/Models/user.ts
export class User {
    FirstName: string;
    LastName: string;
    EmployeeId: number;
}

// users: User[] = [
//     {
//       FirstName: "Uthaya Kumar",
//       LastName: "Ganesan",
//       EmployeeId: 3049571
//     },
//     {
//       FirstName: "Anil Kumar",
//       LastName: "Gupta",
//       EmployeeId: 1234567
//     },
//     {
//       FirstName: "Rajiv Tiwary",
//       LastName: "Malhotra",
//       EmployeeId: 8475757
//     },
//     {
//       FirstName: "Sunil Sinha",
//       LastName: "Venkatesh",
//       EmployeeId: 3049571
//     },
//     {
//       FirstName: "Mani",
//       LastName: "Krishna",
//       EmployeeId: 9003483
//     }
//   ];